fix bugs

+++ main.py
@@ -9,9 +9,9 @@ import requests
 async def get_coords():
     while True:
         try:
-            request = ""
-            response = requests.get(request)
-            yield response["latitude"], response["longitude"]
+            request = "http://192.168.10.101:5000/gps"
+            response = requests.get(request).json()
+            yield response["coordinates"]["latitude"], response["coordinates"]["longitude"]
             await asyncio.sleep(1)
         except:
             pass
@@ -22,14 +22,16 @@ async def main():
     window.showFullScreen()
     mp = gps.MapParams()
     mp.load_map()
+    window.update_additional_widget(os.getcwd()+"\map.png")
     coords_generator = get_coords()
     while True:
         try:
-            latitude, longitude = next(coords_generator)
-            mp.set_coords(latitude, longitude)
-            mp.load_map()
-            window.update_additional_widget(os.getcwd()+"\map.png")
-            os.remove(os.getcwd()+r"\map.png")
+            latitude, longitude = await coords_generator.__anext_
+            if (latitude != 0) & (longitude != 0):
+                mp.set_coords(latitude, longitude)
+                mp.load_map()
+                window.update_additional_widget(os.getcwd()+"\map.png")
+                os.remove(os.getcwd()+r"\map.png")
         except:
             pass
         finally:
